requirements.txt
requests==2.28.1
asgiref==3.5.2
distlib==0.3.6
dj-database-url==1.0.0
django==4.1.1
dnspython==2.2.1
filelock==3.8.0
psycopg2==2.9.3
platformdirs==2.5.2
python-snappy==0.6.1
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
virtualenv==20.16.5
whitenoise==6.2.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
python-environ==0.4.54
boto3~=1.26.13
botocore~=1.29.13